backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
sqlalchemy2-stubs==0.0.2a38
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.2.1
pydantic==2.12.3
pydantic-settings==2.6.0

